+++ packages/form-schema/src/index.ts
@@ -1,6 +1,7 @@
 export { default as serialize } from './serialize';
 export { default as parse } from './parse';
 export { default as mapFieldErrors } from './mapFieldErrors';
+export { validateBeforeSubmit } from './validateBeforeSubmit';
 
 export type {
   GetFieldSchema,
@@ -9,6 +10,7 @@ export type {
   CreateGetFieldSchema,
   Serializer,
   Parser,
+  ValidatorBeforeSubmit,
   ErrorsMapper,
   FieldType,
 } from './types';